Refactor AlertButton component structure by moving it to a dedicated directory and splitting functionality into AlertButtonEdit and AlertButtonCreate components for improved clarity and maintainability in alert management.
